scripts/legacy/requirements_old.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
dj-database-url==2.1.0
    # via -r requirements.in
django==5.2.6
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-cors-headers
django-allauth==0.63.3
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
iniconfig==2.1.0
    # via pytest
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pytest==8.2.0
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via dj-database-url
whitenoise==6.6.0
    # via -r requirements.in
